requirements.txt
slack-bolt==1.27.0
apscheduler==3.11.2
gspread==6.2.1
google-auth==2.49.1
python-dotenv==1.2.1
psycopg2-binary==2.9.11
anthropic==0.88.0
google-api-python-client==2.193.0
